% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitpages.R
\name{KnitPost}
\alias{KnitPost}
\title{Knit Post}
\usage{
KnitPost(input, outfile, figsfolder, cachefolder, base.url = "/")
}
\arguments{
\item{input}{input .Rmd}

\item{outfile}{output .md}

\item{figsfolder}{where figures will be}

\item{cachefolder}{idk}

\item{base.url}{the base directory}
}
\description{
This function converts .Rmd files in a directory to .md files in another directory
}
\keyword{knit}
